toggle subject notifications
