+++ notifications/src/main/java/dk/shape/games/notifications/utils/EnumUtils.kt
@@ -1,4 +1,8 @@
 package dk.shape.games.notifications.utils
 
+import java.util.*
+
 inline fun <reified T : Enum<*>> enumValueOrNull(name: String): T? =
-    T::class.java.enumConstants.firstOrNull { it.name == name }
+    T::class.java.enumConstants.firstOrNull { enumValue ->
+        enumValue.name.toLowerCase(Locale.ROOT) == name
+    }
